Replace a common talk form template with a component
Similar to InterviewInputsFactory, TalkInputsFactory is not using a generated factory/interface because I couldn't get it working with the name service that should be passed to TalkInputs, was getting "Service '' is not defined." because of some mapping applied strangely when only one create() parameter was present probably.
