+++ site/app/Admin/Presenters/templates/Talks/default.latte
@@ -12,7 +12,7 @@
 <div id="pridat-prednasku">
 	<p><a href="#pridat-prednasku" class="open-container block">Přidat přednášku</a></p>
 	<div id="pridat-prednasku-container" class="hidden">
-	{include "common/talkForm.latte", formName: addTalk, talk: null}
+	{control addTalkInputs}
 	</div>
 </div>
 <hr>
