+++ site/app/Admin/Presenters/templates/Talks/talk.latte
@@ -5,5 +5,5 @@
 {/define}
 
 {define #content}
-{include "common/talkForm.latte", formName: editTalk}
+{control editTalkInputs}
 {/define}
