+++ site/config/services.neon
@@ -82,6 +82,7 @@ services:
 	- MichalSpacekCz\Pulse\Sites(@database.pulse.context)
 	- MichalSpacekCz\Tags\Tags
 	- MichalSpacekCz\Talks\TalkFactory(videoFactory: @talkVideoFactory)
+	- MichalSpacekCz\Talks\TalkInputsFactory(videoThumbnails: @talkVideoThumbnails)
 	- MichalSpacekCz\Talks\Talks
 	- MichalSpacekCz\Talks\TalkSlides
 	- MichalSpacekCz\Templating\Filters
